promoted conversation_memory into permamem
